Add editable text and CSS breakpoint at 700px
Text is editable in headers and body of all cards. Saves dynamically and live. One interesting interraction I had to implement was the function setEndofContenteditable, which will always move the cursor to the end of the contenteditable text. This is to counterract the default behavior of it moving back to the beginning of the editable text after every dynamic save of the text. This is avoidable if we use a button to save the new text, but I believe the awkwardness this brings is lesser than the awkwardness of a button in our UI.